+++ src/Symfony/Component/Form/Extension/DataCollector/FormDataExtractor.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\Form\Extension\DataCollector;
 
+use Symfony\Component\Form\DataClassType;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormView;
 use Symfony\Component\Validator\ConstraintViolationInterface;
@@ -24,10 +25,12 @@ class FormDataExtractor implements FormDataExtractorInterface
 {
     public function extractConfiguration(FormInterface $form): array
     {
+        $innerType = $form->getConfig()->getType()->getInnerType();
+
         $data = [
             'id' => $this->buildId($form),
             'name' => $form->getName(),
-            'type_class' => $form->getConfig()->getType()->getInnerType()::class,
+            'type_class' => $innerType instanceof DataClassType ? $innerType->getDataClass() : $innerType::class,
             'synchronized' => $form->isSynchronized(),
             'passed_options' => [],
             'resolved_options' => [],