+++ src/Symfony/Component/Form/Tests/Fixtures/AsFormType/VehicleData.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Form\Tests\Fixtures\AsFormType;
+
+use Symfony\Component\Form\Attribute\AsFormType;
+use Symfony\Component\Form\Attribute\FormField;
+
+#[AsFormType]
+abstract class VehicleData
+{
+    #[FormField]
+    public ?string $name = null;
+}